Better conditional includes, bsd compatability + reworks.
Instead of what I assume was a gnu make specific feature of `ifeq` the `include config.mk` uses `-include` and will not crash if config.mk is not present.
Externized hash_table_realloc and rewrote comments for the functions, renamed the members of `struct hash_table_home` to be more simple.
Added a sandbox executable which is built and for developers, probaly should remove it from the all rule but it is small enough.